Update according to review
Makes total sense to use only one node for the output. But in
combination with having not enough hidden notes it worked
better with four :D The mentioned rule of thumb input +
output / 2 would lead to roughly 25 nodes, which is not
better than the also mentioned 8. Adding only some more than
8 leads to worse accuracies.